Move CLIDebugger out of lib/debug/index
New home: lib/debug/cli
